export support for blackbox tests
So it can be used for plugin development.
